+++ src/components/SearchBar.types.tsx
