+++ src/components/SearchResults.tsx
